Refactor

+++ samples/SvelteKitSample/Program.cs
@@ -4,7 +4,7 @@ using SpaCliMiddleware;
 
 var builder = WebApplication.CreateBuilder(args);
 
-builder.Services.AddControllers();
+// builder.Services.AddControllers();
 
 // Add services to the container.
 builder.Services.AddRazorPages();
@@ -41,23 +41,26 @@ if (hostEnvironment != null)
         app.UseNodejsService(hostEnvironment);
         app.UseMiddleware<NodejsMiddleware>();
     }
+
+    app.UseEndpoints(endpoints => {
+        // endpoints.MapControllers();
+        
+        if (hostEnvironment.IsDevelopment())
+        {
+            // Note: only use spacliproxy in development. 
+            // Production should use "UseSpaStaticFiles()"
+            endpoints.MapToSpaCliProxy(
+                "{*path}",
+                new SpaOptions { SourcePath = "Scripts" },
+                npmScript: "dev",
+                port: /*default(int)*/ 8019, // Allow vite to find own port
+                regex: "SvelteKit v",
+                forceKill: true, // kill anything running on our webpack port
+                useProxy: true, // proxy webpack requests back through our aspnet server
+                runner: ScriptRunnerType.Npm
+            );
+        }
+    });
 }
-
-app.UseEndpoints(endpoints => {
-    endpoints.MapControllers();
-
-    // Note: only use spacliproxy in development. 
-    // Production should use "UseSpaStaticFiles()"
-    endpoints.MapToSpaCliProxy(
-            "{*path}",
-            new SpaOptions { SourcePath = "Scripts" },
-            npmScript: hostEnvironment.IsDevelopment() ? "dev" : "",
-            port: /*default(int)*/ 8019, // Allow vite to find own port
-            regex: "SvelteKit v",
-            forceKill: true, // kill anything running on our webpack port
-            useProxy: true, // proxy webpack requests back through our aspnet server
-            runner: ScriptRunnerType.Npm
-        );
-});
 
 app.Run();
